Move UpdatesTab code into its own class.
