+++ workers/unity/Packages/io.improbable.gdk.debug/NetStatsViewer/NetStatsUpdatesTab.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa51d77d4a61bbe4e89050083051618d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
