+++ workers/unity/Packages/io.improbable.gdk.debug/NetStatsViewer/Templates/NetStatsWindow.uss
@@ -26,3 +26,7 @@
 .component {
     min-width: 160px;
 }
+
+.tab-selector-container {
+    flex-direction: row;
+}
