+++ workers/unity/Packages/io.improbable.gdk.debug/NetStatsViewer/Templates/NetStatsWindow.uxml
@@ -1,8 +1,14 @@
-<UXML xmlns="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements">
+<UXML xmlns="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements" xmlns:ns="Improbable.Gdk.Debug.NetStats">
     <uie:ToolbarMenu name="worldSelector" class="world-selector">
         <Style path="Packages/io.improbable.gdk.debug/NetStatsViewer/Templates/NetStatsWindow.uss" />
     </uie:ToolbarMenu>
-    <VisualElement name="updates" class="tab">
+    <uie:Toolbar name="tabSelectorContainer" class="tab-selector-container">
+        <Style path="Packages/io.improbable.gdk.debug/NetStatsViewer/Templates/NetStatsWindow.uss" />
+        <uie:ToolbarButton name="updateSelector" text="Updates" class="tab-toggle" />
+        <uie:ToolbarButton name="commandSelector" text="Commands" class="tab-toggle" />
+        <uie:ToolbarButton name="worldCommandSelector" text="World Commands" class="tab-toggle" />
+    </uie:Toolbar>
+    <ns:NetStatsUpdatesTab name="updatesTab" class="tab">
         <Style path="Packages/io.improbable.gdk.debug/NetStatsViewer/Templates/NetStatsWindow.uss" />
         <VisualElement name="headerBar" class="header">
             <uie:ToolbarButton text="Component" name="ComponentButton" class="column component" />
@@ -13,7 +19,7 @@
             <VisualElement style="min-width: 15px;" />
         </VisualElement>
         <ScrollView name="updatesContainer" class="updates-view" focusable="false" show-vertical-scroller="true" />
-    </VisualElement>
-    <VisualElement name="commands" class="tab" style="display: none;" />
-    <VisualElement name="worldCommands" class="tab" style="display: none;" />
+    </ns:NetStatsUpdatesTab>
+    <VisualElement name="commandsTab" class="tab" style="display: none;" />
+    <VisualElement name="worldCommandsTab" class="tab" style="display: none;" />
 </UXML>
